# backend/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Prabh Musik API
  description: Production-hardened REST API contract for the Prabh Musik Beat Marketplace.
  version: 1.0.0
servers:
  - url: http://localhost:5005
    description: Local development environment
paths:
...
  /api/users:
    post:
      summary: Create Raw User Profile
      description: Inserts a raw user record into the users table. Used by internal sync processors.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - name
              properties:
                email:
                  type: string
                  format: email
                name:
                  type: string
      responses:
        '201':
          description: Profile created successfully
  /api/beats:
    get:
      summary: List Storefront Catalog Beats
      description: Retrieves published catalog items with optional filtering and sorting. Limit 100 requests per 15 minutes.
      parameters:
        - name: genre
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Catalog beats payload
...
  /api/orders:
    post:
      summary: Create Order Record
      description: Registers a new pending order transaction for beat items.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - customerId
                - beatIds
                - paymentMethod
              properties:
                customerId:
                  type: integer
                beatIds:
                  type: array
                  items:
                    type: integer
                paymentMethod:
                  type: string
                  example: card
      responses:
        '201':
          description: Pending order created successfully
        '400':
          description: Invalid request parameters
        '404':
          description: Customer or beat not found
...
  /api/payments/webhook:
    post:
      summary: Stripe Webhook Entry
      description: Authoritative payment confirmation callback point for Stripe. Enforces Stripe signature verification.
      parameters:
        - name: stripe-signature
          in: header
          required: true
          schema:
            type: string
            description: Stripe webhook signature header
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Webhook event processed or skipped successfully
        '400':
          description: Missing signature or signature verification failed
        '500':
          description: Internal webhook processing failure
  /api/ownerships:
    get:
      summary: List Logged-In Customer's Licensed Library
      description: Returns paginated and sorted beat ownership records for the authenticated user.
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [purchaseDate, beatTitle, price]
            default: purchaseDate
        - name: order
          in: query
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Customer library payload
        '401':
          description: Unauthorized session
...
  /api/uploads/audio:
    post:
      summary: Upload Audio Asset (Admin)
      description: Uploads a beat audio or audio preview file. Allowed formats are mp3, wav, ogg, flac. Max size 25MB.
      security:
        - BearerAuth: []
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [audio, preview]
            default: audio
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                audio:
                  type: string
                  format: binary
      responses:
        '201':
          description: Upload succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResult'
        '400':
          description: Validation failure
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (Admin only)
  /api/uploads/image:
    post:
      summary: Upload Image Asset (Admin)
      description: Uploads a cover image or avatar. Allowed formats are jpg, jpeg, png, webp. Max size 5MB (avatar 2MB).
      security:
        - BearerAuth: []
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [cover, avatar]
            default: cover
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '201':
          description: Upload succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResult'
        '400':
          description: Validation failure
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (Admin only)
  /api/uploads/banner:
    post:
      summary: Upload Banner Image (Admin)
      description: Uploads a channel/beat banner image. Allowed formats are jpg, jpeg, png, webp. Max size 5MB.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                banner:
                  type: string
                  format: binary
      responses:
        '201':
          description: Upload succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResult'
        '400':
          description: Validation failure
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (Admin only)
  /api/uploads/{publicId}:
    patch:
      summary: Replace Uploaded Asset (Admin)
      description: Replaces an existing storage asset atomically. The old asset is only deleted after the new one is verified and written.
      security:
        - BearerAuth: []
      parameters:
        - name: publicId
          in: path
          required: true
          schema:
            type: string
            example: upl_01k3xyz
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Replacement succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResult'
        '400':
          description: Validation failure
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (Admin only)
        '404':
          description: Upload metadata not found or inactive

    delete:
      summary: Delete Asset (Admin)
      description: Soft deletes the metadata record and removes the physical file from the storage provider.
      security:
        - BearerAuth: []
      parameters:
        - name: publicId
          in: path
          required: true
          schema:
            type: string
            example: upl_01k3xyz
      responses:
        '200':
          description: Asset successfully deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (Admin only)
        '404':
          description: Upload metadata not found

    get:
      summary: Get Upload Metadata & Public Link (Admin)
      description: Retrieves the asset database metadata record and constructs a signed public CDN/preview URL.
      security:
        - BearerAuth: []
      parameters:
        - name: publicId
          in: path
          required: true
          schema:
            type: string
            example: upl_01k3xyz
      responses:
        '200':
          description: Metadata successfully returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/UploadResultDetail'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (Admin only)
        '404':
          description: Upload metadata not found
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UploadResult:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/UploadResultDetail'
    UploadResultDetail:
      type: object
      properties:
        id:
          type: integer
          example: 1
        public_id:
          type: string
          example: upl_01k3xyz
        storage_key:
          type: string
          example: audio/2026/08/01/01k3xyz.wav
        asset_type:
          type: string
          example: audio
        mime_type:
          type: string
          example: audio/wav
        file_size:
          type: integer
          example: 1548293
        checksum:
          type: string
          example: a7b5928f9d0c...
        duration:
          type: number
          format: float
          example: 180.5
        image_width:
          type: integer
          nullable: true
          example: null
        image_height:
          type: integer
          nullable: true
          example: null
        uploaded_by:
          type: integer
          example: 2
        status:
          type: string
          example: active
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        publicUrl:
          type: string
          format: uri
          example: https://prabh-musik-assets.r2.cloudflarestorage.com/audio/2026/08/01/01k3xyz.wav?X-Amz-Signature=...